% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSFMnet Package Functions.R
\docType{data}
\name{diff.gene.mat}
\alias{diff.gene.mat}
\title{Example gene expression matrix}
\format{
A matrix with 50 rows and 200 columns
}
\usage{
diff.gene.mat
}
\description{
An example simulated dataset consisting of gene expression counts for G = 50 genes (rows)
across N = 200 cells (columns). The first 100 columns correspond to the control group
and the remaining 100 columns correspond to the treatment group
}
\examples{
data(diff.gene.mat)
rownames(diff.gene.mat)
colnames(diff.gene.mat)

}
\keyword{datasets}
