% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSFMnet Package Functions.R
\docType{data}
\name{mmc.gene.mat}
\alias{mmc.gene.mat}
\title{Mouse microglia cell (MMC) scRNA-seq data}
\format{
A matrix with 98 rows (genes) and 944 columns (cells)
}
\usage{
mmc.gene.mat
}
\description{
Subset of data obtained from the GEO database under accession number GSE90975 and contains the gene expressions from single-cell
analysis of neurodegeneration in microglia cells of mice (Tay et al., 2018). This data includes N = 944 cells G = 98 differentially
expressed genes taken from Figure S1 of the Tay et al. manuscript.
}
\examples{
data(mmc.gene.mat)
rownames(mmc.gene.mat)
colnames(mmc.gene.mat)

}
\references{
Tay, T. L., Dautzenberg, J., Grün, D., & Prinz, M. (2018). Unique microglia recovery population revealed by single-cell RNA-seq following
neurodegeneration. Acta neuropathologica communications, 6(1), 1-11.
}
\keyword{datasets}
