% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSFMnet-package.R
\docType{package}
\name{scSFMnet-package}
\alias{scSFMnet-package}
\alias{scSFMnet}
\title{The 'scSFMnet' package.}
\description{
R package with functions that implement the methodology defined in the
"Single-cell Differential Network Analysis with Sparse Bayesian Factor Models" manuscript.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
