% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSFMnet Package Functions 2.R
\name{sfm.hmc}
\alias{sfm.hmc}
\title{HMC parameter estimation wrapper function for differential network analysis}
\usage{
sfm.hmc(
  Y,
  treatGroup,
  Fac,
  mod.type = "SHS",
  chains = 4,
  n.cores = 1,
  seed = 123,
  ...
)
}
\arguments{
\item{Y}{data.frame or matrix of gene expression counts where the rows correspond to genes and
columns correspond to cells; Y must contain integers and have row names}

\item{treatGroup}{vector or factor of length ncol(Y) indicating the
treatment assignments (control or treatment) of the cells.}

\item{Fac}{number of factors to consider in the model; only a single number is accepted}

\item{mod.type}{type of model to implement; choose one of either "SHS" for Single HorseShoe or
"DHS" for Double HorseShoe}

\item{chains}{a positive integer specifying the number of Markov chains. The rstan default is 4.}

\item{n.cores}{number of cores to use when executing the stan chains in parallel}

\item{seed}{seed for random number generation in rstan}

\item{...}{additional parameters to be passed to sampling() function in rstan}
}
\value{
sfm.stan-class object containing:

\itemize{
\item{Y}{: data.frame or matrix of gene expression counts}
\item{Fac}{: number of factors considered in the model}
\item{stan.fit}{: an object of S4 class stanfit representing the fitted results}
\item{corr.d}{: symmetric estimated differential correlation matrix generated from optimization}
\item{CI.low.d}{: symmetric matrix representing the lower bound of the 95\% credible intervals (CrI) generated from HMC}
\item{CI.upp.d}{: symmetric matrix representing the upper bound of the 95\% credible intervals (CrI) generated from HMC}
\item{CI.eval.d}{: symmetric logical matrix generated from 95\% credible intervals (CrI) from HMC;
TRUE means 95\% CrI does not include 0 (i.e., significant correlation)}
\item{p.val.d}{: symmetric matrix consisting of approximate "p-values" generated from HMC for differential network analysis}
\item{corr.0}{: symmetric estimated correlation matrix generated from optimization for control group}
\item{CI.low.0}{: symmetric matrix representing the lower bound of the 95\% credible intervals (CrI) generated from HMC}
\item{CI.upp.0}{: symmetric matrix representing the upper bound of the 95\% credible intervals (CrI) generated from HMC}
\item{CI.eval.0}{: symmetric logical matrix generated from 95\% credible intervals (CrI) from HMC;
TRUE means 95\% CrI does not include 0 (i.e., significant correlation)}
\item{p.val.0}{: symmetric matrix consisting of approximate "p-values" generated from HMC for control group}
\item{corr.1}{: symmetric estimated correlation matrix generated from optimization for treatment group}
\item{CI.low.1}{: symmetric matrix representing the lower bound of the 95\% credible intervals (CrI) generated from HMC}
\item{CI.upp.1}{: symmetric matrix representing the upper bound of the 95\% credible intervals (CrI) generated from HMC}
\item{CI.eval.1}{: symmetric logical matrix generated from 95\% credible intervals (CrI) from HMC;
TRUE means 95\% CrI does not include 0 (i.e., significant correlation)}
\item{p.val.1}{: symmetric matrix consisting of approximate "p-values" generated from HMC for treatment group}
}
}
\description{
Wrapper function used to implement the Hamiltonian Monte Carlo (HMC) version of the hierarchical
Bayesian factor model defined in the "Single-cell Differential Network Analysis with Sparse Bayesian
Factor Models" manuscript.
}
\examples{
\dontrun{
## Load dataset
data("diff.gene.mat")

## First 100 columns are the control group, the next 100 columns are the treatment group
t_i = rep(0:1,each=100)

## Run HMC via stan
res1 = sfm.hmc(Y = diff.gene.mat, treatGroup = t_i, Fac = 7, mod.type = "SHS", seed = 123)

## Output differential network results
summary(res1)

## Output control network correlation
controlCorr(res1)

## Output treatment network correlation
treatCorr(res1)

}

}
