% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSFMnet Package Functions.R
\name{sfm}
\alias{sfm}
\title{Optimization parameter estimation for differential network analysis}
\usage{
sfm(
  Y,
  treatGroup,
  Fac,
  mod.type = "SHS",
  boot.iter = 1,
  ori.data = FALSE,
  seed = 123,
  stan.iter = 1e+05,
  stan.history.size = 10,
  ...
)
}
\arguments{
\item{Y}{data.frame or matrix of gene expression counts where the rows correspond to genes and
columns correspond to cells; Y must contain integers and have row names}

\item{treatGroup}{vector or factor of length ncol(Y) indicating the
treatment assignments (control or treatment) of the cells.}

\item{Fac}{number of factors to consider in the model; only a single number is accepted}

\item{mod.type}{type of model to implement; choose one of either "SHS" for Single HorseShoe or
"DHS" for Double HorseShoe}

\item{boot.iter}{total number of bootstrap iterations}

\item{ori.data}{if TRUE, the first iteration will consist of the original (non-sampled) data; to be used
when interested in non-bootstrapped results or in the first optimization when bootstrapping}

\item{seed}{seed for random number generation}

\item{stan.iter}{maximum number of optimization iterations}

\item{stan.history.size}{number of update vectors to use in Hessian approximations; for more details refer to
the Stan manual}

\item{...}{additional parameters to be passed to optimize() function in rstan}
}
\value{
sfm.fit-class object containing:

\itemize{
\item{Y}{: data.frame or matrix of gene expression counts}
\item{Fac}{: number of factors considered in the model}
\item{corr.est.0}{: array of estimated correlation matrices for the control group}
\item{corr.est.1}{: array of estimated correlation matrices for the treatment group}
\item{corr.est.d}{: array of estimated correlation difference (control - treatment) matrices}
}
}
\description{
Function used to implement the hierarchical Bayesian factor model defined in the
"Single-cell Differential Network Analysis with Sparse Bayesian Factor Models"
manuscript.
}
\examples{
\dontrun{
## Load dataset
data("diff.gene.mat")

## First 100 columns are the control group, the next 100 columns are the treatment group
t_i = rep(0:1,each=100)

## For first set of bootstrap optimizations set ori.data = TRUE to optimize on original data in first iteration
res1 = sfm(Y = diff.gene.mat, treatGroup = t_i, Fac = 7, mod.type = "SHS", boot.iter = 5, ori.data = TRUE, seed = 123,
stan.iter = 100000, stan.history.size = 10)

}

}
