% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSFMnet Package Functions.R
\name{sfm.results}
\alias{sfm.results}
\title{Estimate the gene-gene correlation matrices from bootstrapped samples}
\usage{
sfm.results(sfm.list, ori.samp = 1)
}
\arguments{
\item{sfm.list}{list where each element contains an sfm.fit-class object; each element of the list
contains an object from a different parallel bootstrap chain}

\item{ori.samp}{location of list element that contains optimization on non-sampled data; default is 1}
}
\value{
sfm.corr-class object containing:

\itemize{
\item{corr.d}{: symmetric estimated differential correlation matrix generated from optimization}
\item{CI.low.d}{: symmetric matrix representing the lower bound of the 95\% confidence intervals (CI) generated from the bootstrap optimizations for differential network analysis}
\item{CI.upp.d}{: symmetric matrix representing the upper bound of the 95\% confidence intervals (CI) generated from the bootstrap optimizations for differential network analysis}
\item{CI.eval.d}{: symmetric logical matrix generated from 95\% confidence intervals (CI) from bootstrap optimizations for differential network analysis;
TRUE means 95\% CI does not include 0 (i.e., significant correlation)}
\item{p.val.d}{: symmetric matrix consisting of approximate "p-values" generated from bootstrap optimizations for differential network analysis}
\item{corr.0}{: symmetric estimated correlation matrix generated from optimization for control group}
\item{CI.low.0}{: symmetric matrix representing the lower bound of the 95\% confidence intervals (CI) generated from the bootstrap optimizations for control group}
\item{CI.upp.0}{: symmetric matrix representing the upper bound of the 95\% confidence intervals (CI) generated from the bootstrap optimizations for control group}
\item{CI.eval.0}{: symmetric logical matrix generated from 95\% confidence intervals (CI) from bootstrap optimizations for control group;
TRUE means 95\% CI does not include 0 (i.e., significant correlation)}
\item{p.val.0}{: symmetric matrix consisting of approximate "p-values" generated from bootstrap optimizations for control group}
\item{corr.1}{: symmetric estimated correlation matrix generated from optimization for treatment group}
\item{CI.low.1}{: symmetric matrix representing the lower bound of the 95\% confidence intervals (CI) generated from the bootstrap optimizations for treatment group}
\item{CI.upp.1}{: symmetric matrix representing the upper bound of the 95\% confidence intervals (CI) generated from the bootstrap optimizations for treatment group}
\item{CI.eval.1}{: symmetric logical matrix generated from 95\% confidence intervals (CI) from bootstrap optimizations for treatment group;
TRUE means 95\% CI does not include 0 (i.e., significant correlation)}
\item{p.val.1}{: symmetric matrix consisting of approximate "p-values" generated from bootstrap optimizations for treatment group}
}
}
\description{
Function used to combine and analyze bootstrapped results from one or more sets of
correlation samples generated by the \code{sfm} function.
}
\details{
The correlation estimates are determined from original optimization on original (non-sampled) data for each gene-gene pair.

To determine whether the correlation is significant, a 95\% confidence interval (CI) is determined from the bootstrap samples. If the CI
includes 0, the correlation is deemed to be non-significant and the "CI.eval" element is FALSE. If the CI does not include 0, the correlation is
considered significant and the "CI.eval" element is TRUE.

An alternative measurement of significance is the approximate "p-value" calculation, which is output in the "p.val" matrix. This "p-value" is determined
by finding the smallest "a" value such that the 100(1-a)\% CI contains 0. The corresponding "a" value represents the proportion of posterior distribution
that is outside the smallest confidence interval that contains 0.

In most cases a significant 95\% CI from "CI.eval" will correspond with an approximate "p-value" < 0.05 from "p.val".
}
\examples{
\dontrun{
## Load dataset
data("diff.gene.mat")

## First 100 columns are the control group, the next 100 columns are the treatment group
t_i = rep(0:1,each=100)

## For first set of bootstrap optimizations set ori.data = TRUE to optimize on original data in first iteration
res1 = sfm(Y = diff.gene.mat, treatGroup = t_i, Fac = 7, mod.type = "SHS", boot.iter = 5, ori.data = TRUE, seed = 123,
stan.iter = 100000, stan.history.size = 10)

## For other sets of bootstrap optimizations set ori.data = FALSE to optimize on resampled data
res2 = sfm(Y = diff.gene.mat, treatGroup = t_i, Fac = 7, mod.type = "SHS", boot.iter = 5, ori.data = FALSE, seed = 1234,
stan.iter = 100000, stan.history.size = 10)

## Combine bootstrap results from multiple sfm.fit-class objects
## Note: The first object has the optimization estimates from original data hence ori.samp = 1
res.list = list(res1,res2)
boot.res = sfm.results(res.list, ori.samp=1)

## Output differential network results
summary(boot.res)

## Output control network correlation
controlCorr(boot.res)

## Output treatment network correlation
treatCorr(boot.res)

}

}
